requirements.txt
gradio
tensorflow-cpu==2.12.0
numpy
pillow
pandas
matplotlib